First and most important: Istore Deleted. logic: we open a session per-user per-store, it will be IStoreService.
since we go to DL-based model, Class store hold only actions about one store that done by one user.
therfore, StoreService will go away and Store will implement IStoreService (and will add the "open store" fucntion as an external fucntion)

interface IStoreService got all the functions of Istore

SystemDL - minor changes.
ModuleGlobalHandler + IStoreService - remove ProductNotFound Enum value, return exception insted

class store - minor changes.
